refactor: remove hooks folders
