+++ apps/editor/messages/es.po
@@ -760,81 +760,6 @@ msgctxt "QGLi0O"
 msgid "URL address"
 msgstr "Dirección URL"
 
-#: src/hooks/use-category-labels.ts
-msgctxt "djJp6c"
-msgid "History"
-msgstr "Historia"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "dPgwrL"
-msgid "Math"
-msgstr "Matemáticas"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "fFt3il"
-msgid "Arts"
-msgstr "Artes"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "GsBRWL"
-msgid "Languages"
-msgstr "Idiomas"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "hlmkcL"
-msgid "Health"
-msgstr "Salud"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "hwL7a0"
-msgid "Engineering"
-msgstr "Ingeniería"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "I1ZCPK"
-msgid "Technology"
-msgstr "Tecnología"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "KOuDOE"
-msgid "Economics"
-msgstr "Economía"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "mn6pIm"
-msgid "Culture"
-msgstr "Cultura"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "n5D/Ku"
-msgid "Communication"
-msgstr "Comunicación"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "OV/2+4"
-msgid "Society"
-msgstr "Sociedad"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "qydxOd"
-msgid "Science"
-msgstr "Ciencia"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "u3ovjm"
-msgid "Law"
-msgstr "Derecho"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "w1Fanr"
-msgid "Business"
-msgstr "Negocios"
-
-#: src/hooks/use-category-labels.ts
-msgctxt "YMDBNH"
-msgid "Geography"
-msgstr "Geografía"
-
 #: src/lib/error-messages.ts
 msgctxt "12q6p9"
 msgid "Category not found in this course"
@@ -919,3 +844,78 @@ msgstr "Formato JSON inválido. Por favor revisa tu archivo"
 msgctxt "yVHcXv"
 msgid "Organization not found"
 msgstr "Organización no encontrada"
+
+#: src/lib/use-category-labels.ts
+msgctxt "djJp6c"
+msgid "History"
+msgstr "Historia"
+
+#: src/lib/use-category-labels.ts
+msgctxt "dPgwrL"
+msgid "Math"
+msgstr "Matemáticas"
+
+#: src/lib/use-category-labels.ts
+msgctxt "fFt3il"
+msgid "Arts"
+msgstr "Artes"
+
+#: src/lib/use-category-labels.ts
+msgctxt "GsBRWL"
+msgid "Languages"
+msgstr "Idiomas"
+
+#: src/lib/use-category-labels.ts
+msgctxt "hlmkcL"
+msgid "Health"
+msgstr "Salud"
+
+#: src/lib/use-category-labels.ts
+msgctxt "hwL7a0"
+msgid "Engineering"
+msgstr "Ingeniería"
+
+#: src/lib/use-category-labels.ts
+msgctxt "I1ZCPK"
+msgid "Technology"
+msgstr "Tecnología"
+
+#: src/lib/use-category-labels.ts
+msgctxt "KOuDOE"
+msgid "Economics"
+msgstr "Economía"
+
+#: src/lib/use-category-labels.ts
+msgctxt "mn6pIm"
+msgid "Culture"
+msgstr "Cultura"
+
+#: src/lib/use-category-labels.ts
+msgctxt "n5D/Ku"
+msgid "Communication"
+msgstr "Comunicación"
+
+#: src/lib/use-category-labels.ts
+msgctxt "OV/2+4"
+msgid "Society"
+msgstr "Sociedad"
+
+#: src/lib/use-category-labels.ts
+msgctxt "qydxOd"
+msgid "Science"
+msgstr "Ciencia"
+
+#: src/lib/use-category-labels.ts
+msgctxt "u3ovjm"
+msgid "Law"
+msgstr "Derecho"
+
+#: src/lib/use-category-labels.ts
+msgctxt "w1Fanr"
+msgid "Business"
+msgstr "Negocios"
+
+#: src/lib/use-category-labels.ts
+msgctxt "YMDBNH"
+msgid "Geography"
+msgstr "Geografía"
