+++ apps/editor/src/components/slug-editor.tsx
@@ -11,7 +11,7 @@ import { Check, CircleAlert, Link as LinkIcon, X } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useExtracted } from "next-intl";
 import { useState, useTransition } from "react";
-import { useSlugCheck } from "@/hooks/use-slug-check";
+import { useSlugCheck } from "@/lib/use-slug-check";
 
 export function SlugEditorSkeleton() {
   return <Skeleton className="mx-4 h-5 w-32" />;
