+++ apps/main/messages/en.po
@@ -1002,96 +1002,6 @@ msgctxt "eUzs3M"
 msgid "Send feedback, questions, or suggestions to us. Fill in the form below or email us directly at hello@zoonk.com."
 msgstr "Send feedback, questions, or suggestions to us. Fill in the form below or email us directly at hello@zoonk.com."
 
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "djJp6c"
-msgid "History"
-msgstr "History"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "dPgwrL"
-msgid "Math"
-msgstr "Math"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "fFt3il"
-msgid "Arts"
-msgstr "Arts"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "GsBRWL"
-msgid "Languages"
-msgstr "Languages"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "hlmkcL"
-msgid "Health"
-msgstr "Health"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "hwL7a0"
-msgid "Engineering"
-msgstr "Engineering"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "I1ZCPK"
-msgid "Technology"
-msgstr "Technology"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "KOuDOE"
-msgid "Economics"
-msgstr "Economics"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "mn6pIm"
-msgid "Culture"
-msgstr "Culture"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "n5D/Ku"
-msgid "Communication"
-msgstr "Communication"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "OV/2+4"
-msgid "Society"
-msgstr "Society"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "qydxOd"
-msgid "Science"
-msgstr "Science"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "u3ovjm"
-msgid "Law"
-msgstr "Law"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "w1Fanr"
-msgid "Business"
-msgstr "Business"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "YMDBNH"
-msgid "Geography"
-msgstr "Geography"
-
 #: src/lib/activities.ts
 msgctxt "1TOMnj"
 msgid "Listening"
@@ -1218,9 +1128,69 @@ msgid "{category} Courses"
 msgstr "{category} Courses"
 
 #: src/lib/categories.ts
+msgctxt "djJp6c"
+msgid "History"
+msgstr "History"
+
+#: src/lib/categories.ts
+msgctxt "dPgwrL"
+msgid "Math"
+msgstr "Math"
+
+#: src/lib/categories.ts
 msgctxt "EMi23I"
 msgid "Explore all {category} courses"
 msgstr "Explore all {category} courses"
+
+#: src/lib/categories.ts
+msgctxt "fFt3il"
+msgid "Arts"
+msgstr "Arts"
+
+#: src/lib/categories.ts
+msgctxt "GsBRWL"
+msgid "Languages"
+msgstr "Languages"
+
+#: src/lib/categories.ts
+msgctxt "hlmkcL"
+msgid "Health"
+msgstr "Health"
+
+#: src/lib/categories.ts
+msgctxt "hwL7a0"
+msgid "Engineering"
+msgstr "Engineering"
+
+#: src/lib/categories.ts
+msgctxt "I1ZCPK"
+msgid "Technology"
+msgstr "Technology"
+
+#: src/lib/categories.ts
+msgctxt "KOuDOE"
+msgid "Economics"
+msgstr "Economics"
+
+#: src/lib/categories.ts
+msgctxt "mn6pIm"
+msgid "Culture"
+msgstr "Culture"
+
+#: src/lib/categories.ts
+msgctxt "n5D/Ku"
+msgid "Communication"
+msgstr "Communication"
+
+#: src/lib/categories.ts
+msgctxt "OV/2+4"
+msgid "Society"
+msgstr "Society"
+
+#: src/lib/categories.ts
+msgctxt "qydxOd"
+msgid "Science"
+msgstr "Science"
 
 #: src/lib/categories.ts
 msgctxt "Rz9oL+"
@@ -1228,6 +1198,21 @@ msgid "{category} courses"
 msgstr "{category} courses"
 
 #: src/lib/categories.ts
+msgctxt "u3ovjm"
+msgid "Law"
+msgstr "Law"
+
+#: src/lib/categories.ts
+msgctxt "w1Fanr"
+msgid "Business"
+msgstr "Business"
+
+#: src/lib/categories.ts
 msgctxt "XL8IWT"
 msgid "{category} courses. The best {category} online courses, interactive lessons to learn."
 msgstr "{category} courses. The best {category} online courses, interactive lessons to learn."
+
+#: src/lib/categories.ts
+msgctxt "YMDBNH"
+msgid "Geography"
+msgstr "Geography"
