+++ apps/main/messages/pt.po
@@ -1002,96 +1002,6 @@ msgctxt "eUzs3M"
 msgid "Send feedback, questions, or suggestions to us. Fill in the form below or email us directly at hello@zoonk.com."
 msgstr "Envie feedback, perguntas ou sugestões para a gente. Preencha o formulário abaixo ou envie um email diretamente para contato@zoonk.com."
 
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "djJp6c"
-msgid "History"
-msgstr "História"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "dPgwrL"
-msgid "Math"
-msgstr "Matemática"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "fFt3il"
-msgid "Arts"
-msgstr "Artes"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "GsBRWL"
-msgid "Languages"
-msgstr "Idiomas"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "hlmkcL"
-msgid "Health"
-msgstr "Saúde"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "hwL7a0"
-msgid "Engineering"
-msgstr "Engenharia"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "I1ZCPK"
-msgid "Technology"
-msgstr "Tecnologia"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "KOuDOE"
-msgid "Economics"
-msgstr "Economia"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "mn6pIm"
-msgid "Culture"
-msgstr "Cultura"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "n5D/Ku"
-msgid "Communication"
-msgstr "Comunicação"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "OV/2+4"
-msgid "Society"
-msgstr "Sociedade"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "qydxOd"
-msgid "Science"
-msgstr "Ciências"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "u3ovjm"
-msgid "Law"
-msgstr "Direito"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "w1Fanr"
-msgid "Business"
-msgstr "Negócios"
-
-#: src/hooks/use-categories.ts
-#: src/lib/categories.ts
-msgctxt "YMDBNH"
-msgid "Geography"
-msgstr "Geografia"
-
 #: src/lib/activities.ts
 msgctxt "1TOMnj"
 msgid "Listening"
@@ -1218,9 +1128,69 @@ msgid "{category} Courses"
 msgstr "Cursos de {category}"
 
 #: src/lib/categories.ts
+msgctxt "djJp6c"
+msgid "History"
+msgstr "História"
+
+#: src/lib/categories.ts
+msgctxt "dPgwrL"
+msgid "Math"
+msgstr "Matemática"
+
+#: src/lib/categories.ts
 msgctxt "EMi23I"
 msgid "Explore all {category} courses"
 msgstr "Explorar todos os cursos de {category}"
+
+#: src/lib/categories.ts
+msgctxt "fFt3il"
+msgid "Arts"
+msgstr "Artes"
+
+#: src/lib/categories.ts
+msgctxt "GsBRWL"
+msgid "Languages"
+msgstr "Idiomas"
+
+#: src/lib/categories.ts
+msgctxt "hlmkcL"
+msgid "Health"
+msgstr "Saúde"
+
+#: src/lib/categories.ts
+msgctxt "hwL7a0"
+msgid "Engineering"
+msgstr "Engenharia"
+
+#: src/lib/categories.ts
+msgctxt "I1ZCPK"
+msgid "Technology"
+msgstr "Tecnologia"
+
+#: src/lib/categories.ts
+msgctxt "KOuDOE"
+msgid "Economics"
+msgstr "Economia"
+
+#: src/lib/categories.ts
+msgctxt "mn6pIm"
+msgid "Culture"
+msgstr "Cultura"
+
+#: src/lib/categories.ts
+msgctxt "n5D/Ku"
+msgid "Communication"
+msgstr "Comunicação"
+
+#: src/lib/categories.ts
+msgctxt "OV/2+4"
+msgid "Society"
+msgstr "Sociedade"
+
+#: src/lib/categories.ts
+msgctxt "qydxOd"
+msgid "Science"
+msgstr "Ciências"
 
 #: src/lib/categories.ts
 msgctxt "Rz9oL+"
@@ -1228,6 +1198,21 @@ msgid "{category} courses"
 msgstr "Cursos de {category}"
 
 #: src/lib/categories.ts
+msgctxt "u3ovjm"
+msgid "Law"
+msgstr "Direito"
+
+#: src/lib/categories.ts
+msgctxt "w1Fanr"
+msgid "Business"
+msgstr "Negócios"
+
+#: src/lib/categories.ts
 msgctxt "XL8IWT"
 msgid "{category} courses. The best {category} online courses, interactive lessons to learn."
 msgstr "Cursos de {category}. Os melhores cursos online de {category}, aulas interativas pra aprender."
+
+#: src/lib/categories.ts
+msgctxt "YMDBNH"
+msgid "Geography"
+msgstr "Geografia"
