+++ apps/main/src/app/[locale]/(catalog)/courses/category-pills.tsx
@@ -4,8 +4,8 @@ import { buttonVariants } from "@zoonk/ui/components/button";
 import { Skeleton } from "@zoonk/ui/components/skeleton";
 import { useSelectedLayoutSegment } from "next/navigation";
 import { useExtracted } from "next-intl";
-import { useCategories } from "@/hooks/use-categories";
 import { Link } from "@/i18n/navigation";
+import { useCategories } from "@/lib/categories";
 
 export function CategoryPillsSkeleton() {
   return (
