+++ packages/core/package.json
@@ -18,7 +18,6 @@
   "exports": {
     "./ai": "./src/ai.ts",
     "./alternative-titles/add": "./src/alternative-titles/add-alternative-titles.ts",
-    "./alternative-titles/find": "./src/alternative-titles/find-alternative-title.ts",
     "./auth": "./src/auth/auth.ts",
     "./auth/client": "./src/auth/client.ts",
     "./auth/hooks/auth-state": "./src/auth/hooks/auth-state.ts",
